add authorization

+++ util/user_id.go
@@ -9,14 +9,14 @@ var (
 	userIdKey = "user_id"
 )
 
-func ContextWithUserID(ctx context.Context, userID int64) context.Context {
+func ContextWithUserID(ctx context.Context, userID uint) context.Context {
 	return context.WithValue(ctx, userIdKey, userID)
 }
 
-func RequestWithUserID(r *http.Request, userID int64) *http.Request {
+func RequestWithUserID(r *http.Request, userID uint) *http.Request {
 	return r.WithContext(ContextWithUserID(r.Context(), userID))
 }
 
-func UserIDFromContext(ctx context.Context) int64 {
-	return ctx.Value("user_id").(int64)
+func UserIDFromContext(ctx context.Context) uint {
+	return ctx.Value("user_id").(uint)
 }
